post delete, edit 안되는 버그 수정
- delete는 prisma 테이블에 관계를 설정하면서 생긴 문제라 관계 테이블에 onDelete와 onUpdate 생겼을 시 어떻게 할지 알려줌
- post edit은 useQuery로 가져온 데이터로 form의 state에 설정해줬어야했는데 useQuery가 값을 가져오기 전에 useState의 값을 설정해서 생기는 문제였음. 그래서 컴포넌트를 분리해서 useQuery로 값을 가져오면 form을 보여주도록 설정
